chore: split tinder service
